+++ .codesandbox/templates/vanilla/src/index.ts
@@ -7,6 +7,8 @@ const canvas = (window.canvas = new fabric.Canvas(el, {
   width: 500,
   height: 500,
   preserveObjectStacking: true,
+  allowTouchScrolling: true,
+  enablePointerEvents: true,
 }));
 
 testCase(canvas);