+++ src/EventTypeDefs.ts
@@ -265,7 +265,9 @@ type TPointerEvents<Prefix extends string> = Record<
   > &
   Record<`${Prefix}wheel`, TPointerEventInfo<WheelEvent>> &
   Record<`${Prefix}over`, TPointerEventInfo & InEvent> &
-  Record<`${Prefix}out`, TPointerEventInfo & OutEvent>;
+  Record<`${Prefix}out`, TPointerEventInfo & OutEvent> &
+  Record<'pinch', TPointerEventInfo & { scale: number }> &
+  Record<'rotate', TPointerEventInfo & { rotation: number }>;
 
 export type TPointerEventNames =
   | WithBeforeSuffix<'down'>
@@ -295,7 +297,6 @@ export interface ObjectEvents
   deselected: Partial<TEvent> & {
     target: FabricObject;
   };
-
   // tree
   added: { target: Group | Canvas | StaticCanvas };
   removed: { target: Group | Canvas | StaticCanvas };